Moved DescriptorLayoutBuilder into separate files
