+++ Engine/src/Platform/Vulkan/VulkanDescriptors.h
@@ -2,41 +2,8 @@
 #include <span>
 #include <vulkan/vulkan_core.h>
 
-#include "VkBootstrap.h"
-
 namespace Engine
 {
-    class DescriptorLayoutBuilder
-    {
-    public:
-        explicit DescriptorLayoutBuilder(vkb::Device const& logicalDevice);
-
-        DescriptorLayoutBuilder& SetStageFlags(VkShaderStageFlags flags);
-
-        DescriptorLayoutBuilder& SetCreateFlags(VkDescriptorSetLayoutCreateFlags flags);
-
-        DescriptorLayoutBuilder& AddBinding(uint32_t binding, VkDescriptorType type);
-
-        DescriptorLayoutBuilder& ClearBindings();
-
-        DescriptorLayoutBuilder& SetAllocationCallbacks(VkAllocationCallbacks* callbacks);
-
-        VkDescriptorSetLayout Build();
-
-    private:
-        struct LayoutInfo
-        {
-            VkDevice LogicalDevice = VK_NULL_HANDLE;
-
-            VkShaderStageFlags StageFlags;
-            std::vector<VkDescriptorSetLayoutBinding> Bindings;
-
-            VkDescriptorSetLayoutCreateFlags CreateFlags = 0;
-
-            VkAllocationCallbacks* AllocationCallbacks = VK_NULL_HANDLE;
-        } _info;
-    };
-
     class DescriptorAllocator
     {
     public:
@@ -63,6 +30,6 @@ namespace Engine
         std::vector<PoolSizeRatio> _ratios;
         std::vector<VkDescriptorPool> _fullPools;
         std::vector<VkDescriptorPool> _readyPools;
-        uint32_t _setsPerPool;
+        uint32_t _setsPerPool = 0;
     };
 }
